Create RecipeList and RecipeBox Components
The 2 components inside the App component are the RecipeList
component, which contains many RecipeBox component's. A static list
of recipes is passed as props to RecipeList. A random Wikipedia
image of Carbonara is used for all recipes (obviously this will be
changed).

This is simply a static version; state will be introduced in the future.