Move gitignore to Root Directory
The .gitignore file was originally created inside the client folder. I'm
moving it up one directory, since there are also files in the /api
directory that should be ignored.

I also updated the contents of .gitignore to ignore the node_modules
folder in both api and client, as well as update the location of build/
to ignore.