Create Initial API GET Endpoint with Dummy Data
An API endpoint of GET /api/recipes was created on the Express app
which returns hard coded json data. In the future this will connect
to AWS DynamoDB, but for now hard coding the data is good enough.

On the front end, state was added to the RecipeList component so that
the component would re-render once the fetch call was returned from
the API.

Additionally, a 'proxy_pass' variable was set in the client's
package.json. This is so the front end will proxy any unknown requests
to localhost:5000 (aka any time an API request is sent, it will go to
the Express app). This is a workaround for CORS, and is only to be used
in development.