Add support for using Docker
I decided to learn Docker for this project and utilize it for
creating development & production environments. docker-compose.yml
will be used for production, and will be combined with
docker-compose.override.yml for development.

In development, there will be a 3 conainer setup. One container each for
the API, react app, and web server. The web server isn't really needed
for this setup, but it can't be removed from the docker-compose file.

In production, there will only be containers for the API and web server.
The web server will serve the static React build, and proxy any /api
requests to the API container.