Get coming-soon image from S3 Bucket
I decided to store the iamges for this project in AWS S3 as a way to
learn more about the AWS ecosystem.

A GET request to /api/images/:key gets the signed url from S3
associated with that key and returns it to the client. If that key
doesn't exist the API returns 404 and if there is some other error
with getting the signed url, the API returns 500.

On the client side, for each recipe returned from DynamoDB, a request is
made to the API for the signed url.
TODO: Add the S3 image key to DynamoDB and use provide that to the API
to get the recipe specific image.