Use CloudFront CDN to Serve Images
Seeing as making calls to S3 for each image is expensive (both in terms
of $$ and time), I setup CloudFront to serve images from the S3 bucket.
CloudFront will cache the images and store each image at a static url,
so that no API call needs to be made from the front end. If an image
updates (i.e the picture changes but the name stays the same), it will
still be available at the same URL.

In the API this means that the /images/:key route is redundant along
with any associated middleware functions.

As mentioned before, on the client no API call is necessary and the
static base URL in conjunction with the image name from DynamoDB can be
used to link the CloudFront image source.