Add Testing for the API
All API's should have tests to ensure that it has the desired
behaviour. Here I decided to use Mocha and Chai as my testing and
assertion libraries, since they seem easy to setup and use.

Since this API is extremely simple (only has a single GET endpoint)
the number of tests that can be written are limited. I only test that
the endpoint will return an array of objects, each containing a set of
keys, and that a 404 error will be returned on a bad endpoint.

The package*.json files are updated to reflect the new dependencies
installed, and the actual app.js needs to be exported so it can be used
by the tests.