Fix Production Build Styling Bug
When the React production build was created and served inside the nginx
container, there were a few styles that weren't being applied.
Specifically, the text styles and centering were absent.

Upon extensive searching, I discovered that the CSS was being
interpreted as a stylesheet, but served as text/plain instead of css.
Upon further research, I found that this was due to nginx not setting
the proper mime types within the configuration. Apparently when setting
up my nginx.conf, I ripped out all of the default blocks, including a
line that included the proper mime types. By adding this line back, the
error with styles in the production build went away.