Create GitHub Action Workflow to Unit Test API
I have made the decision to use GitHub Actions as my choice for CI/CD.
I didn't want to use a large system like Jenkins of TeamCity, so my
options were narrowed down to either CircleCI or GitHub Actions. I went
with GitHub Actions because I already use GitHub and using this for
CI/CD means one less application I need to look after. I did some
research and found that CircleCI and GitHub Actions offer many of the
same services for simple applications like this.

Getting started with GitHub Actions required creating a YAML file in a
new .github/workflows directory (I added the API subdirectory to break
up any future tests I'll create). This file instructs GitHub to only run
the action on pushes and pull requests that invlude the /api directory.
It will install the dependencies and then run the tests provided.

Updating the npm script was necessary because there is a bug in Mocha
that causes it to not exit after the tests are complete.
See this for more info:
https://stackoverflow.com/questions/50372866/mocha-not-exiting-after-test