Rename Workflow File
The workflow file will be used for all steps (not just API unit tests)
until GitHub Actions support more complex dependencies for workflows.
The reason for this is that the integration and deploy jobs should
depend on the unit tests, which is not currently possible. In order to
have all jobs in the same workflow file, adding if statements to the
integration and deploy jobs is necessary.

See here for more info:
https://github.community/t/how-do-i-specify-job-dependency-running-in-another-workflow/16482