Add Tests & Deploy Jobs to Workflow
All tests as well as the deploy job are being moved to a single workflow
due to GitHub Actions not allowing workflows to depend on each other
(see previous commit for more info).

This commit creates dummy jobs for the Client/Integration tests and
Deploy. These will be updated in the future once the actual tests/deploy
flow have been created.