Fix Bug With Deploy Job in Workflow File (#13)
The deploy job had a dependency on the integration job, but the
integration job was only setup to run on pull request events. This means
that it would never run on pushses to master branch, and this caused the
deploy job to never run. To fix that I updated the integration job to
check for pull requests or pushes to master branch.

There was also a typo when checking the ref on the deploy job.