Create the Integration Test
As of now, simply setting up the Docker Compose containers is enough for
an integration test. In the future, this requirement may change, but as
of now this should be sufficient.

The main components required for the integration test is to checkout the
code, setup the environment variables (for the API service), build the
client code and copy to the web service, and then running
docker-compose.