Create AWS EC2 Instance Terraform
In order to create an EC2 instance, a security group with inbound
and outbound rules must be created. Cloud-Init is used to run the
commands necessary to install Docker on a Linux AMI.

Additionally, I created an Elastic IP so that each time the instance
is destroyed and re-created it will have the same IP (assuming that
the EIP resource doesn't get destroyed).