Add Build Stage to Workflow
As part of implementing CI/CD, I need a way to automatically build and
push docker images to Docker Hub. Using the official Docker build-push
action, I'm able to provide inputs that will build the API and Web
images, and push them to Docker Hub with the appropriate tags.

1 tag will be a snapshot of that individual build and will have the calver based
tag in the form yyyy.mm.dd-SHA7 so that there is a history of all the
versions. The other tag will be latest, and this is what will be pulled
down in production.