Read AWS Secrets from .tfvars
Including AWS keys in version control is a horrible idea for many
reasons. Using the functionality of Terraform's .tfvars file allows me
to put the secrets in that file and hide it from version control. This
way when Terraform runs locally, it will run as expected, all while hiding the
secrets from Git.

Additionally, a .tfvars.example file is included to show the format of
the .tfvars file without actually exposing the real secrets.