Create ActiveTag and ActiveTagList Components
Create the 2 components that will be used for rendering tags.
Additionally, create a RecipeApp container component that holds
everything except the title of the app.

The active tags are passed in as props to the lower components, and can
be deleted by clicking the delete button.

The alogrithm to filter the recipes based on the tags still needs to be
created.