Update the Displayed Recipes Based on Tags
In order for implementing tags to have any value, the recipes themselves
need to be updated when the user selects a tag. The initial tags are
still hardcoded, but they will filter the recipes that are shown and the
user can click on the X to remove a specific tag.

To Do:
- set no tags initially and add a list of tags to be selected
- improve spacing of tags
- create client side, tag specific tests

Note: If a recipe in the DB doesn't have a tags key, this will cause an
error in the client (because you can't determine if a key is in a null
object). At one point I had some logic to handle this, but it was
slightly messy and this will force all recipes to be added without
missing information.