Fix Bug When Setting Environment Variables
Using the set-env command has been deprecated in Actions for security
reasons. This is causing the build section of workflow.yml to fail.

GitHub now provides the functionality to add envionment vars to
the $GITHUB_ENV file in the form of "key=value" and then access the
variable in a later step with $key.

https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-commands-for-github-actions#environment-files