Add /login Route and Ability to POST a Recipe
In order to add new recipes a user must first be able to login. The
login route takes a username and password and checks the Users table
to see if there are any entries that match. If so, the user is returned
a JWT that expires in 1h (this time is subject to change). If the username
or password are incorrect, the request is given an unsuccessful status
code.

If a user wants to add a new recipe, 2 middleware functions were
added to help with this. First there is a function that takes the
request and verifies the token in the Authorization header is valid.
Second, a function checks that a recipe doesn't already exist with
the name in the request. If both functions pass, then the recipe is
added to the DB.