Update Node Version for API Image
Somehow I was using a pretty old version of node (node 8) in the API
docker container. This was probably because I saw some tutorial or StackOverflow
article using it and I followed suit. When I started using promises, this
caused some issues with the new endpoints I added.

After some troubleshooting and debugging the JS code, I bumped the
image version and it worked.

Node 14.15 was chosen, since that is the current LTS version available.